requirements.txt
PyQt6>=6.6.0
pysstv>=0.5.4
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
sounddevice>=0.4.6
moviepy>=1.0.3
